project1_facial_key_points_detection/requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
altair=1.2.1=py_0
asn1crypto=0.22.0=py36_0
atari-py=0.1.7=pypi_0
atomicwrites=1.3.0=pypi_0
attrs=19.1.0=py_0
audioread=2.1.6=py36_0
av=0.3.3=py36_2
awscli=1.16.17=py36_0
backcall=0.1.0=py_0
backports=1.0=py36_1
backports.functools_lru_cache=1.4=py36_1
backports.weakref=1.0rc1=py36_1
beautifulsoup4=4.6.0=py36_0
binutils_impl_linux-64=2.31.1=h6176602_1
binutils_linux-64=2.31.1=h6176602_7
blas=1.1=openblas
bleach=1.5.0=py36_0
blinker=1.4=py_0
bokeh=0.12.13=py36_0
boto=2.48.0=py36h6e4cd66_1
boto3=1.9.7=py_0
botocore=1.12.7=py_0
box2d=2.3.2=pypi_0
box2d-kengz=2.3.3=pypi_0
bresenham=0.2=pypi_0
bz2file=0.98=py36_0
bzip2=1.0.6=1
ca-certificates=2019.11.28=hecc5488_0
cairo=1.14.10=h58b644b_4
cassandra-driver=3.11.0=pypi_0
certifi=2019.11.28=py36_0
cffi=1.11.2=py36_0
chardet=3.0.4=py36_0
clang_variant=1.0=cling_6.14.06
clangdev=5.0.0=h935a590_1004
click=6.7=py_1
cling=0.5=he860b03_1006
cloudpickle=0.2.2=py36_4
colorama=0.3.9=py_1
colour=0.1.5=py_0
conda=4.6.14=py36_0
cppzmq=4.3.0=h1b96802_1001
cryptography=2.1.4=py36_0
cryptopp=7.0.0=hf484d3e_1003
cudatoolkit=8.0=3
cudnn=7.0.5=cuda8.0_0
curl=7.49.0=1
cycler=0.10.0=py36_0
cython=0.29.7=pypi_0
cytoolz=0.9.0.1=py36h14c3975_1001
dask-core=0.16.1=py_0
dbus=1.10.22=0
decorator=4.0.11=py36_0
defusedxml=0.5.0=py_1
dill=0.2.7.1=py36_0
docutils=0.14=py36_0
dronekit=2.9.2=pypi_0
entrypoints=0.2.3=py36_1
et_xmlfile=1.0.1=py36_0
eventlet=0.22.0=py_0
expat=2.2.5=0
fastcache=1.0.2=py36_0
ffmpeg=3.4.1=0
flask=0.12.2=py36_0
flask-socketio=2.9.2=py36_0
fontconfig=2.12.4=h88586e7_1
freetype=2.8.1=0
future=0.16.0=py36_0
gcc_impl_linux-64=7.3.0=habb00fd_1
gcc_linux-64=7.3.0=h553295d_7
gensim=3.4.0=py36_0
geos=3.6.2=1
gettext=0.19.7=1
giflib=5.1.4=0
glfw=1.8.0=pypi_0
glib=2.53.6=h0d2cc06_1
gmp=6.1.2=0
graphite2=1.3.13=hf484d3e_1000
greenlet=0.4.12=py36_0
gst-plugins-base=1.14.0=he9c8b40_0
gstreamer=1.14.0=hb31296c_0
gxx_impl_linux-64=7.3.0=hdf63c60_1
gxx_linux-64=7.3.0=h553295d_7
gym=0.9.6=pypi_0
h5py=2.7.1=py36_1
harfbuzz=1.5.0=h2545bd6_0
hdf5=1.8.18=2
heapdict=1.0.0=py36_0
html2text=2018.1.9=pypi_0
html5lib=0.9999999=py36_0
icu=58.2=0
idna=2.6=py36_1
imageio=2.1.2=py36_0
intel-openmp=2018.0.0=hc7b2577_8
ipykernel=4.9.0=py36_0
ipython=6.5.0=py36_0
ipython-sql=0.3.9=py36_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.0.5=py36_0
itsdangerous=0.24=py_2
jasper=1.900.1=4
jdc=0.0.8=pypi_0
jdcal=1.3=py36_0
jedi=0.10.2=py36_0
jieba=0.39=pypi_0
jinja2=2.10=py36_0
jmespath=0.9.3=py36_0
joblib=0.11=py36_0
jpeg=9b=2
json5=0.8.5=py_0
jsonschema=2.6.0=py36_0
jupyter_client=5.2.4=py_3
jupyter_core=4.4.0=py_0
jupyterlab=1.0.9=py36_0
jupyterlab_server=1.0.0=py_0
keras=2.0.9=py36_0
libblas=3.8.0=7_h6e990d7_netlib
libcblas=3.8.0=7_h6e990d7_netlib
libedit=3.1.20170329=0
libffi=3.2.1=3
libgcc=7.2.0=h69d50b8_2
libgcc-ng=7.3.0=hdf63c60_0
libgfortran=3.0.0=1
libgfortran-ng=7.2.0=h9f7466a_2
libgpuarray=0.7.5=0
libiconv=1.15=h470a237_1
liblapack=3.8.0=7_h6e990d7_netlib
libopenblas=0.2.20=h9ac9557_4
libpng=1.6.34=0
libpq=9.6.6=h1f21990_0
librosa=0.5.1=py36_1
libsodium=1.0.15=1
libstdcxx-ng=7.3.0=hdf63c60_0
libtiff=4.0.9=he6b73bb_1
libtorch=0.1.12=nomkl_0
libuuid=2.32.1=h14c3975_1000
libwebp=0.5.2=7
libxcb=1.13=0
libxml2=2.9.7=0
llvmlite=0.20.0=py36_0
locket=0.2.0=py36_1
lockfile=0.12.2=pypi_0
lxml=4.1.1=pypi_0
mako=1.0.7=pypi_0
markdown=2.6.9=py36_0
markupsafe=1.0=py36_0
matplotlib=2.1.0=py36hba5de38_0
mistune=0.8.3=py_0
mkl=2018.0.1=h19d6760_4
mock=2.0.0=py36_0
monotonic=1.3=pypi_0
more-itertools=7.0.0=pypi_0
moviepy=0.2.3.2=py36_0
mpmath=1.0.0=py_0
msgpack=0.5.6=pypi_0
msgpack-python=0.4.8=py36_0
mujoco-py=1.50.1.59=pypi_0
nbconvert=5.4.0=py36_1
nbformat=4.4.0=py36_0
nbzip=0.1.0=pypi_0
nccl=1.3.4=cuda8.0_1
ncurses=5.9=10
networkx=1.11=py36_0
ninja=1.8.2=1
nlohmann_json=3.4.0=hf484d3e_0
nltk=3.2.5=py_0
nomkl=1.0=0
notebook=5.7.0=py36_0
numba=0.35.0=np112py36_0
numexpr=2.6.4=py36_nomklh2c28f9d_0
numpy=1.12.1=py36_blas_openblas_200
oauthlib=2.0.6=py_0
olefile=0.44=py36_0
openblas=0.2.19=2
opencv=3.3.1=py36h0f6f1c3_0
openpyxl=2.5.0b1=py36_0
openssl=1.0.2u=h516909a_0
packaging=16.8=py36_0
pandas=0.23.3=py36_0
pandoc=1.19.2=0
pandocfilters=1.4.1=py36_0
parso=0.4.0=py_0
partd=0.3.8=py36_0
patsy=0.4.1=py36_0
pbr=3.1.1=py36_0
pcre=8.39=0
pexpect=4.3.1=py36_0
pickleshare=0.7.4=py36_0
pillow=5.2.0=py36_0
pinyin=0.4.0=pypi_0
pip=18.1=py36_0
pixman=0.34.0=1
plotly=2.0.15=py36_0
pluggy=0.11.0=pypi_0
pomegranate=0.9.0=py36h3010b51_0
prettytable=0.7.2=py_2
prometheus_client=0.3.1=py_1
prompt_toolkit=1.0.15=py36_0
protobuf=3.5.1=py36_3
psutil=5.4.0=py36_0
psycopg2=2.7.4=py36hb7f436b_0
ptyprocess=0.5.2=py36_0
py=1.8.0=pypi_0
py4j=0.10.7=py36_0
pyasn1=0.4.4=py_1
pycosat=0.6.3=py36_0
pycparser=2.18=py36_0
pycurl=7.43.0=py36_0
pydot=1.2.4=py_0
pydotplus=2.0.2=py36_0
pyglet=1.3.0=pypi_0
pygments=2.2.0=py36_0
pygpu=0.7.5=py36_0
pyjwt=1.5.3=py_0
pymavlink=2.2.20=pypi_0
pymc3=3.2=py36_0
pyopengl=3.1.0=pypi_0
pyopenssl=17.4.0=py36_0
pyparsing=2.2.0=py36_0
pyqt=5.6.0=py36_4
pyrsistent=0.15.2=py36h516909a_0
pysocks=1.6.8=py36_1
pyspark=2.4.3=py_0
pytest=4.5.0=pypi_0
python=3.6.3=3
python-crfsuite=0.9.2=py36_0
python-dateutil=2.6.1=py36_0
python-engineio=2.0.2=py_0
python-socketio=1.8.4=py_0
python-speech-features=0.6=pypi_0
pytorch=0.4.0=py36_cuda8.0.61_cudnn7.1.2_1
pytz=2017.3=py_2
pywavelets=0.5.2=py36_1
pyyaml=3.12=py36_1
pyzmq=17.1.2=py36hae99301_0
qt=5.6.2=7
readline=7.0=hb321a52_4
requests=2.18.4=py36_1
requests-oauthlib=0.8.0=py36_1
requests-toolbelt=0.9.1=pypi_0
resampy=0.2.0=py36_1
rsa=3.4.2=py_1
ruamel_yaml=0.11.14=py36_0
s3transfer=0.1.13=py36_0
scikit-image=0.14.2=py36he6710b0_0
scikit-learn=0.19.1=py36_blas_openblas_200
scipy=1.2.1=py36h09a28d5_1
seaborn=0.8.1=py36_0
seekwell=0.1=pypi_0
send2trash=1.5.0=py_0
setuptools=38.4.0=py36_0
shapely=1.6.4=py36_0
simplegeneric=0.8.1=py36_0
simplewebsocketserver=0.1.1=pypi_0
sip=4.18=py36_1
six=1.11.0=py36_1
smart_open=1.5.6=py36_1
snownlp=0.12.3=pypi_0
soundfile=0.9.0.post1=pypi_0
sqlalchemy=1.1.13=py36_0
sqlite=3.20.1=h6d8b0f3_1
sqlparse=0.3.0=py_0
statsmodels=0.8.0=py36_0
sympy=1.0=py36_0
tensorflow=1.3.0=py36_0
terminado=0.8.1=py36_0
testpath=0.3.1=py36_0
theano=1.0.1=py36_1
tk=8.6.7=0
toolz=0.8.2=py_2
torchvision=0.2.1=py36_1
tornado=4.5.3=py36_0
tqdm=4.11.2=py36_0
traitlets=4.3.2=py36_0
tweepy=3.5.0=py36_0
twython=3.6.0=py36_0
udacity-pa=0.2.9=pypi_0
urllib3=1.22=py36_0
util-linux=2.21=0
utm=0.4.2=py36_0
uvloop=0.8.1=py36_0
vega=0.4.4=py36_1
vincent=0.4.4=py36_0
wcwidth=0.1.7=py36_0
webencodings=0.5=py36_0
websockets=4.0.1=py36_0
werkzeug=0.14.1=py_0
wheel=0.30.0=py36_2
widgetsnbextension=3.1.0=py36_0
wordcloud=1.4.1=py36_0
wptools=0.4.7=pypi_0
x264=20131217=3
xeus=0.17.0=h6e680fc_1000
xeus-cling=0.4.9=0
xlrd=1.1.0=py_2
xorg-libxau=1.0.8=3
xorg-libxdmcp=1.1.2=3
xtl=0.5.4=h6bb024c_0
xz=5.2.3=0
yaml=0.1.6=0
zeromq=4.2.5=hfc679d8_4
zlib=1.2.11=0
